+++ CHANGELOG.md
@@ -66,10 +66,10 @@ All notable changes to the "UnrealVsCodeHelper" extension will be documented in 
 
 ## `[0.3.0] - 2022/05/28`
 ### Added
-- `DocExplorer:` New command ***UVCH.OpenUnrealDoc*** who's opening documentation page in your browser
-- `DocExplorer:` New shortcut ***alt+f1*** Triggering ***UVCH.OpenUnrealDoc*** with the current selection has keyword
+- `DocExplorer:` New command ***UVCH.QuickSearch*** who's opening documentation page in your browser
+- `DocExplorer:` New shortcut ***alt+f1*** Triggering ***UVCH.QuickSearch*** with the current selection has keyword
 - `DocExplorer:` New interface panel ***UVCH Browser***, a small browser for unreal documention INSIDE VsCode
-- `DocExplorer:` New shortcut ***alt+shift+f1*** triggering ***UVCH.UnrealSearch*** Allowing you do to do a reseach from the current selection without opening the browser tab
+- `DocExplorer:` New shortcut ***alt+shift+f1*** triggering ***UVCH.BrowserSearch*** Allowing you do to do a reseach from the current selection without opening the browser tab
 
 ## `[0.2.0] - 2022/05/27`
 ### Added